concatenated two files
